@@ -5,9 +5,9 @@
   "google_analytics": "UA-78609319-4",
   "private": false,
   "dependencies": {
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.58",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/icons": "^4.11.3",
+    "@material-ui/lab": "^4.0.0-alpha.61",
     "@types/googlemaps": "^3.43.3",
     "@types/markerclustererplus": "^2.1.33",
     "@types/react": "^17.0.11",
@@ -21,10 +21,10 @@
     "lodash": "^4.17.21",
     "perfect-scrollbar": "^1.5.1",
     "query-string": "^7.0.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-chartist": "^0.14.4",
     "react-document-title": "^2.0.3",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-ga": "^3.3.0",
     "react-google-maps": "^9.4.5",
     "react-router-dom": "^5.2.0",
@@ -39,23 +39,23 @@
     "deploy": "webpack --display-error-details --mode=production --config webpack.prod.js"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.5",
-    "@babel/plugin-proposal-class-properties": "^7.14.5",
-    "@babel/plugin-proposal-decorators": "^7.14.5",
-    "@babel/plugin-proposal-export-namespace-from": "^7.14.5",
-    "@babel/plugin-proposal-function-sent": "^7.14.5",
-    "@babel/plugin-proposal-json-strings": "^7.14.5",
-    "@babel/plugin-proposal-numeric-separator": "^7.14.5",
-    "@babel/plugin-proposal-object-rest-spread": "^7.14.5",
-    "@babel/plugin-proposal-throw-expressions": "^7.14.5",
+    "@babel/core": "^7.18.10",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-proposal-decorators": "^7.18.10",
+    "@babel/plugin-proposal-export-namespace-from": "^7.18.9",
+    "@babel/plugin-proposal-function-sent": "^7.18.6",
+    "@babel/plugin-proposal-json-strings": "^7.18.6",
+    "@babel/plugin-proposal-numeric-separator": "^7.18.6",
+    "@babel/plugin-proposal-object-rest-spread": "^7.18.9",
+    "@babel/plugin-proposal-throw-expressions": "^7.18.6",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "@babel/plugin-syntax-import-meta": "^7.10.4",
-    "@babel/plugin-transform-async-to-generator": "^7.14.5",
-    "@babel/plugin-transform-runtime": "^7.14.5",
-    "@babel/plugin-transform-spread": "^7.14.5",
-    "@babel/preset-env": "^7.14.5",
-    "@babel/preset-react": "^7.14.5",
-    "@babel/runtime": "^7.14.5",
+    "@babel/plugin-transform-async-to-generator": "^7.18.6",
+    "@babel/plugin-transform-runtime": "^7.18.10",
+    "@babel/plugin-transform-spread": "^7.18.9",
+    "@babel/preset-env": "^7.18.10",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/runtime": "^7.18.9",
     "babel-loader": "^8.2.2",
     "babel-plugin-import": "^1.13.3",
     "clean-webpack-plugin": "^4.0.0-alpha.0",
