@@ -9,22 +9,43 @@ import withStyles from "@material-ui/core/styles/withStyles";
 import Button from "@material-ui/core/Button";
 
 import buttonStyle from "assets/jss/material-dashboard-react/components/buttonStyle.jsx";
-function RegularButton({ ...props }) {
-  const {
-    classes,
-    color,
-    round,
-    children,
-    disabled,
-    simple,
-    size,
-    block,
-    link,
-    justIcon,
-    className,
-    muiClasses,
-    ...rest
-  } = props;
+
+// function RegularButton({ ...props }) {
+//   const {
+//     classes,
+//     color,
+//     round,
+//     children,
+//     disabled,
+//     simple,
+//     size,
+//     block,
+//     link,
+//     justIcon,
+//     className,
+//     muiClasses,
+//     ...rest
+//   } = props;
+//   const btnClasses = classNames({
+//     [classes.button]: true,
+//     [classes[size]]: size,
+//     [classes[color]]: color,
+//     [classes.round]: round,
+//     [classes.disabled]: disabled,
+//     [classes.simple]: simple,
+//     [classes.block]: block,
+//     [classes.link]: link,
+//     [classes.justIcon]: justIcon,
+//     [className]: className
+//   });
+//   return (
+//     <Button {...rest} classes={muiClasses} className={btnClasses}>
+//       {children}
+//     </Button>
+//   );
+// }
+
+const RegularButton = React.forwardRef(({ classes, color, round, children, disabled, simple, size, block, link, justIcon, className, muiClasses, ...rest }, ref) => {
   const btnClasses = classNames({
     [classes.button]: true,
     [classes[size]]: size,
@@ -38,11 +59,11 @@ function RegularButton({ ...props }) {
     [className]: className
   });
   return (
-    <Button {...rest} classes={muiClasses} className={btnClasses}>
+    <Button {...rest} classes={muiClasses} className={btnClasses} ref={ref}>
       {children}
     </Button>
   );
-}
+});
 
 RegularButton.propTypes = {
   classes: PropTypes.object.isRequired,
