@@ -71,7 +71,7 @@ class HeaderLinks extends React.Component {
         </Button>
         <div className={classes.manager}>
           <Button
-            buttonRef={node => {
+            ref={node => {
               this.anchorEl = node;
             }}
             color={window.innerWidth > 959 ? "transparent" : "white"}
